+++ src/us/kbase/test/groups/service/api/GroupsAPITest.java
@@ -262,6 +262,7 @@ public class GroupsAPITest {
 			.with("id", "id2")
 			.with("private", true)
 			.with("role", "None")
+			.with("resources", Collections.emptyMap())
 			.build();
 
 	@Test